+++ packages/odata-service-inquirer/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @sap-ux/odata-service-inquirer
 
+## 0.3.0
+
+### Minor Changes
+
+-   1ea831d5: Adds support for service url prompting
+
+### Patch Changes
+
+-   Updated dependencies [1ea831d5]
+    -   @sap-ux/inquirer-common@0.3.0
+
+## 0.2.3
+
+### Patch Changes
+
+-   Updated dependencies [ad93a484]
+    -   @sap-ux/project-access@1.22.0
+    -   @sap-ux/telemetry@0.4.32
+
 ## 0.2.2
 
 ### Patch Changes