+++ packages/ui5-library-inquirer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/ui5-library-inquirer
 
+## 0.2.15
+
+### Patch Changes
+
+-   Updated dependencies [1ea831d5]
+    -   @sap-ux/inquirer-common@0.3.0
+
 ## 0.2.14
 
 ### Patch Changes